requirements.txt
requests>=2.32.3
beautifulsoup4>=4.13.3
pandas>=2.2.3
mysql-connector-python>=9.2.0